requirements.txt
httpx>=0.25.0
python-dotenv>=1.0.0
mcp>=0.2.0 
bs4>=0.0.1
